feat: Validations in ActivityService

+++ src/main/java/com/limatech/juriprocessos/services/process/ActivityService.java
@@ -3,28 +3,40 @@ package com.limatech.juriprocessos.services.process;
 import com.limatech.juriprocessos.dtos.process.CreateActivityDTO;
 import com.limatech.juriprocessos.exceptions.process.ActivityNotFoundException;
 import com.limatech.juriprocessos.exceptions.process.ProcessNotFoundException;
+import com.limatech.juriprocessos.exceptions.users.ForbiddenActionException;
+import com.limatech.juriprocessos.exceptions.users.UserNotFoundException;
 import com.limatech.juriprocessos.models.process.entity.Activity;
+import com.limatech.juriprocessos.models.process.entity.Group;
 import com.limatech.juriprocessos.models.process.entity.Process;
+import com.limatech.juriprocessos.models.users.entity.User;
 import com.limatech.juriprocessos.repository.process.ActivityRepository;
 import com.limatech.juriprocessos.repository.process.ProcessRepository;
+import com.limatech.juriprocessos.repository.users.UserRepository;
+import com.limatech.juriprocessos.services.interfaces.UserValidation;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.UUID;
 
 @Service
-public class ActivityService {
+public class ActivityService implements UserValidation {
 
     private final ActivityRepository activityRepository;
     private final ProcessRepository processRepository;
+    private final UserRepository userRepository;
 
     @Autowired
-    public ActivityService(ActivityRepository activityRepository, ProcessRepository processRepository) {
+    public ActivityService(ActivityRepository activityRepository, ProcessRepository processRepository, UserRepository userRepository) {
         this.activityRepository = activityRepository;
         this.processRepository = processRepository;
+        this.userRepository = userRepository;
     }
 
     public Activity createActivity(CreateActivityDTO activityDTO) {
+        validateUserPermission(activityDTO);
+
         Process process = processRepository.findById(activityDTO.getProcessId()).orElseThrow(() -> new ProcessNotFoundException("Process not found"));
         activityDTO.setProcess(process);
 
@@ -34,11 +46,14 @@ public class ActivityService {
     }
 
     public void deleteActivity(UUID id) {
+        validateUserPermission(id);
+
         Activity activity = activityRepository.findById(id).orElseThrow(() -> new ActivityNotFoundException("Activity not found"));
         activityRepository.deleteById(id);
     }
 
     public Activity updateActivity(UUID id, CreateActivityDTO activityDTO) {
+        validateUserPermission(id);
 
         Activity activity = activityRepository.findById(id).orElseThrow(() -> new ActivityNotFoundException("Activity not found"));
         if(activityDTO.getName() != null) {
@@ -59,5 +74,62 @@ public class ActivityService {
 
     public Activity getActivity(UUID id) {
         return activityRepository.findById(id).orElseThrow(() -> new ActivityNotFoundException("Activity not found"));
+    }
+
+    @Override
+    public void validateUserPermission(UUID id) {
+        User currentUserFromContext = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        UUID currentUserId = currentUserFromContext.getId();
+        User currentUserFromDB = userRepository.findById(currentUserId).orElseThrow(() -> new UserNotFoundException(
+                "User not found"));
+
+        Activity activity = activityRepository.findById(id).orElseThrow(() -> new ActivityNotFoundException("Activity not found"));
+        Process process = activity.getProcess();
+
+
+        if(
+                !process.getUser().getId().toString().equals(currentUserId.toString()) &&
+                !userCanWriteOnProcess(currentUserFromDB, process) &&
+                !this.isUserAdmin()
+        ) {
+            throw new ForbiddenActionException();
+        }
+    }
+
+    public void validateUserPermission(CreateActivityDTO activityDTO) {
+        User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        UUID currentUserId = currentUser.getId();
+        User currentUserFromDB = userRepository.findById(currentUserId).orElseThrow(() -> new UserNotFoundException(
+                "User not found"));
+
+        Process process = processRepository.findById(activityDTO.getProcessId()).orElseThrow(() -> new ProcessNotFoundException("Process not found"));
+
+        if(
+                !process.getUser().getId().toString().equals(currentUserId.toString()) &&
+                !userCanWriteOnProcess(currentUserFromDB, process) &&
+                !this.isUserAdmin()
+        ) {
+            throw new ForbiddenActionException();
+        }
+    }
+
+    public boolean userCanWriteOnProcess(User user, Process process) {
+        List<Group> groupsThatUserCanWrite = user.getCanWrite();
+        List<UUID> groupsThatUserCanWriteIds = groupsThatUserCanWrite.stream().map(Group::getId).toList();
+
+        List<Group> groupsFromProcess = process.getGroups();
+        List<UUID> groupsFromProcessIds = groupsFromProcess.stream().map(Group::getId).toList();
+
+        if(process.getUser().getId().toString().equals(user.getId().toString())) {
+            return true;
+        }
+
+        for (UUID uuid : groupsThatUserCanWriteIds) {
+            if (groupsFromProcessIds.contains(uuid)) {
+                return true;
+            }
+        }
+
+        return false;
     }
 }
