+++ src/main/java/com/limatech/juriprocessos/services/process/GroupService.java
@@ -5,6 +5,7 @@ import com.limatech.juriprocessos.dtos.process.CreateGroupDTO;
 import com.limatech.juriprocessos.dtos.process.ManageProcessToGroupDTO;
 import com.limatech.juriprocessos.exceptions.process.GroupNotFoundException;
 import com.limatech.juriprocessos.exceptions.process.ProcessNotFoundException;
+import com.limatech.juriprocessos.exceptions.users.ForbiddenActionException;
 import com.limatech.juriprocessos.exceptions.users.UserNotFoundException;
 import com.limatech.juriprocessos.models.process.entity.Group;
 import com.limatech.juriprocessos.models.process.entity.Process;
@@ -12,14 +13,16 @@ import com.limatech.juriprocessos.models.users.entity.User;
 import com.limatech.juriprocessos.repository.process.GroupRepository;
 import com.limatech.juriprocessos.repository.process.ProcessRepository;
 import com.limatech.juriprocessos.repository.users.UserRepository;
+import com.limatech.juriprocessos.services.interfaces.UserValidation;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.UUID;
 
 @Service
-public class GroupService {
+public class GroupService implements UserValidation {
 
     private final GroupRepository groupRepository;
     private final ProcessRepository processRepository;
@@ -34,6 +37,8 @@ public class GroupService {
     }
 
     public Group createGroup(CreateGroupDTO groupDTO) {
+        validateUserPermission(groupDTO);
+
         User user = userRepository.findById(groupDTO.getUserId()).orElseThrow(() -> new UserNotFoundException("User not found"));
         groupDTO.setUser(user);
         Group group = groupDTO.toEntity();
@@ -41,11 +46,15 @@ public class GroupService {
     }
 
     public void deleteGroup(UUID id) {
+        validateUserPermission(id);
+
         Group group  = groupRepository.findById(id).orElseThrow(() -> new GroupNotFoundException("Group not found"));
         groupRepository.deleteById(id);
     }
 
     public Group updateGroup(UUID id, CreateGroupDTO groupDTO) {
+        validateUserPermission(id);
+
         Group group = groupRepository.findById(id).orElseThrow(() -> new GroupNotFoundException("Group not found"));
 
         if(groupDTO.getName() != null)
@@ -59,10 +68,14 @@ public class GroupService {
     }
 
     public Group getGroup(UUID id) {
+        validateUserPermission(id);
+
         return groupRepository.findById(id).orElseThrow(() -> new GroupNotFoundException("Group not found"));
     }
 
     public Group addProcess(UUID id, ManageProcessToGroupDTO manageProcessToGroupDTO) {
+        validateUserPermission(id);
+
         Group group = groupRepository.findById(id).orElseThrow(() -> new GroupNotFoundException("Group not found"));
         Process process = processRepository.findById(manageProcessToGroupDTO.getProcessId()).orElseThrow(() -> new ProcessNotFoundException(
                 "Process not found"));
@@ -73,6 +86,8 @@ public class GroupService {
     }
 
     public Group removeProcess(UUID id, ManageProcessToGroupDTO manageProcessToGroupDTO) {
+        validateUserPermission(id);
+
         Group group = groupRepository.findById(id).orElseThrow(() -> new GroupNotFoundException("Group not found"));
         Process process =
                 processRepository.findById(manageProcessToGroupDTO.getProcessId()).orElseThrow(() -> new ProcessNotFoundException("Process not found"));
@@ -82,6 +97,8 @@ public class GroupService {
     }
 
     public Group addUsers(AddAndRemoveUserToGroupDTO groupDTO) {
+        validateUserPermission(groupDTO);
+
         Group group = groupRepository.findById(groupDTO.getGroupId()).orElseThrow(() -> new GroupNotFoundException("Group " +
                 "not found"));
         List<UUID> userIdsToAddInCanRead = groupDTO.getCanRead();
@@ -97,6 +114,8 @@ public class GroupService {
     }
 
     public Group removeUsers(AddAndRemoveUserToGroupDTO groupDTO) {
+        validateUserPermission(groupDTO);
+
         Group group = groupRepository.findById(groupDTO.getGroupId()).orElseThrow(() -> new GroupNotFoundException("Group " +
                 "not found"));
         List<UUID> userIdsToRemoveInCanRead = groupDTO.getCanRead();
@@ -109,5 +128,36 @@ public class GroupService {
         usersToRemoveInCanWrite.forEach(group::removeCanWrite);
 
         return groupRepository.save(group);
+    }
+
+    @Override
+    public void validateUserPermission(UUID id) {
+        User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        UUID currentUserId = currentUser.getId();
+
+        if (!id.toString().equals(currentUserId.toString()) && !this.isUserAdmin()) {
+            throw new ForbiddenActionException();
+        }
+    }
+
+    private void validateUserPermission(CreateGroupDTO groupDTO) {
+        User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        UUID currentUserId = currentUser.getId();
+
+        if (!groupDTO.getUserId().toString().equals(currentUserId.toString()) && !this.isUserAdmin()) {
+            throw new ForbiddenActionException();
+        }
+    }
+
+    private void validateUserPermission(AddAndRemoveUserToGroupDTO groupDTO) {
+        User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        UUID currentUserId = currentUser.getId();
+
+        Group group = groupRepository.findById(groupDTO.getGroupId()).orElseThrow(() -> new GroupNotFoundException("Group not found"));
+        User userOwnerOfGroup = group.getUser();
+
+        if (!userOwnerOfGroup.getId().toString().equals(currentUserId.toString()) && !this.isUserAdmin()) {
+            throw new ForbiddenActionException();
+        }
     }
 }
